FaST-LMM

SpatialOmicsGPs.fastlmm_fullrank — Function

fastlmm_fullrank(y,K; covariates = [], mean = true)

For a linear mixed model / Gaussian process,

\[y \sim N\bigl(X\beta, \sigma^2(K + \delta I) \bigr)\]

where $y$ is the response vector, $X$ is a matrix of covariates, and $K$ is a full-rank kernel matrix, compute the REMLs of the variance parameter $\sigma^2$ and the variance ratio $\delta$ using FaST-LMM with a full-rank kernel matrix. Compared to the original FaST-LMM algorithm, we first project out the (optional) covariates, incl. an (optional) constant off-set (mean=true), from the response vector and the kernel matrix. This avoids all matrix computations in the variance parameter estimation, but means all variance estimates are restricted maximum-likelihood estimates (REMLs). Estimates for the fixed effects $\beta$ are not computed.

SpatialOmicsGPs.delta_mle_fullrank — Function

delta_mle_fullrank(λ, yr)

Compute the MLE of the variance ratio δ given the eigenvalues of the kernel matrix and the rotated response vector by solving the non-convex optimization problem formulated in the FaST-LMM paper.

SpatialOmicsGPs.minus_log_like_fullrank — Function

minus_log_like_fullrank(δ, λ, yr)

Compute the minus log-likelihood of the model, scaled by the number of samples and without constant factors, given the variance ratio δ, the eigenvalues of the kernel matrix, the rotated response vector and (optionally) the rotated covariates.

SpatialOmicsGPs.sigma2_mle_fullrank — Function

sigma2_mle_fullrank(δ, λ, yr)

Compute the MLE of the variance parameter given the variance ratio δ, the eigenvalues of the kernel matrix, and the rotated response vector.

SpatialOmicsGPs.project_orth_covar — Function

project_orth_covar(y, K, X)

Project out the covariates from the response vector and the kernel matrix.

SpatialOmicsGPs.softplus — Function

softplus(x)

Compute the softplus function, i.e. log(1 + exp(x)), element-wise.